[general] added define DCX_USE_C11 to defines.h that when set to 1 enabled c++11 feature usage. changed most loops using vectors to use ranged for() loops if c++11 is in use. removed some NULL checks before delete. updated windows version defines for Windows 8 + IE v10 + GDI+ 1.1 added COLOR_GLASS to $dcx(GetSystemColor,COLOR_....) to return the colour being used for the glass effect.
[dialog]
changed /xdialog -R +p to use a different for() loop when getting the points.
changed parseCursorType() to use a map object. This is just for testing & will likely be changed lots.
changed DrawCaret() to use a slightly different method of getting a black pen object.

[trayicon]
changed idExists()/modifyIcon()/AddIconId()/DeleteIconId() to take a const id arg.

[statusbar]
changed deletePartInfo() to do a pPart == NULL check before the loop.

[edit]
changed $xdid().text to type numtok() as int.

[list]
changed /xdid -A +f to type max_lines as int.

[tstring]
type'd some unsigned vars to signed.

[dcxmenu]
changed how setMenuBar() copies the vector to a temp vector.

[misc]
changed isRegexMatch() to support C++11 regex.


git-svn-id: https://dcx.googlecode.com/svn/branches/dcxutf@1117 17a06da3-8c20-fa79-550f-e9588e724b77

+++ Classes/mirc/dcxedit.cpp
@@ -217,7 +217,7 @@ void DcxEdit::parseInfoRequest( const TString &input, PTCHAR szReturnValue) cons
 			if (numtok > 3) {
 				const int nLine = input.getnexttok( ).to_int();	// tok 4
 
-				if (nLine > 0 && nLine <= this->m_tsText.numtok(TEXT("\r\n"))) {
+				if (nLine > 0 && nLine <= (int)this->m_tsText.numtok(TEXT("\r\n"))) {
 					lstrcpyn(szReturnValue, this->m_tsText.gettok(nLine, TEXT("\r\n")).to_chr(), MIRC_BUFFER_SIZE_CCH);
 					return;
 				}
